Fix loading toplevel directory from subdir in prelude
The loadAsDirectory function of the prelude script prepended to
browserified modules strips off trailing slashes before concatenating
the argument with '/package.js' and '/index'. If the argument is a
single slash, i.e. a path pointing to the root directory, it should be
preserved when passing to path.resolve subsequently.

Fix this problem by stripping off trailing slashes only for
package/index lookup.