Added comments to enable chopping up of the bundle into seperate files.
I'm not convinced this is the best way to achieve this: outputing the individual parts would be best,
though I'm not sure this is possible.

Secondly, putting the $body of the wrappers on the same line as the starter boiler plate. This is so
line numbers reported at runtime from a file match up with the file that is edited.