Failing test case for --debug --standalone
This demonstrates that with the --standalone option the UMD prelude
is adding a source line to the output. Since browserify doesn't change
the source map generated by browser-pack, which accommodates its require
runtime preamble, the result is a source map with an off-by-one error in
generated line numbers.

Previously, I attempted to offload --standalone onto browser-pack, but it
can be easily handled by just asserting that the additional prelude not
consume an additional source line.