Remove unnecessary workaround to hide full path
This was previously added in 5fe121b to prevent the absolute path to the
"process" module from exposed in the bundle.

Recently a change landed in insert-module-globals [1] that prevents this from
happening in the first place, thus this workaround is no longer needed.

This is a breaking change for those who are calling require("_process")
directly (which should not happen as it was never advertised as a public api).

[1] https://github.com/substack/insert-module-globals/pull/54