add failing test that tries to round-trip dep
Browserify-incremental makes the assumption that we can round-trip `dep` events
(slightly mutated to form an object with `row.file` as the key) back into the
cache with no ill effects.  This is true except for the edge case shown here -
when you are both exposing a module and requiring it in the same bundle.