fix failing test about exposing and requiring the same module
It should be safe to not walk down an exposed module, since it's being taken
care of at the top-level.  If we don't do this, `module-deps` tries to open a
file using the module's exposed name rather than a valid filename, which errors
out.

This is only triggered by passing in a cache value that should be valid.