Upgrade stream-http to v3
This version of stream-http drops support for IE10 and below.

This should not require a major version bump since browserify 14.0.0 already upgraded the buffer package and dropped support for IE8-10. Since stream uses buffer internally, stream is already not working in IE8-10.

This lets us get the latest and best improvements.