fix `require('process')`
5fe121b added a semi-private `_process` alias for the builtin `process`
shim. Uses of the global `process` would be replaced by
`require('_process')`. That works but you could then not use
`require('process')`. As far as I know that has never worked in
browserify.

There was never a reason for modules to use `require('process')` since
the global is available, but it is a thing that exists that should be
supported. For webpack 5 users it would also be helpful if the core
module shims used `require('process')` because they could take
simpler steps to polyfill the environment correctly. And browserify must
support `require('process')` before we can do that.